feat: add mage spell
