+++ VERSION
@@ -1,1 +1,1 @@
-0.3.4
+0.3.7
