+++ data/db/classes.json
@@ -159,7 +159,7 @@
       "y": 2.0
     },
     "spells": [
-      "fireball_wall"
+      "firewall"
     ],
     "children": []
   }
