+++ src/Application/src/GameLogic.cpp
@@ -35,7 +35,6 @@ game::GameLogic::GameLogic(ThePURGE &game) :
     sinkOnGameStarted.connect<&GameLogic::slots_game_start>(*this);
 
     sinkOnPlayerBuyClass.connect<&GameLogic::slots_apply_classes>(*this);
-    sinkOnPlayerBuyClass.connect<&GameLogic::slots_purchase_classes>(*this);
     sinkOnPlayerLevelUp.connect<&GameLogic::slots_level_up>(*this);
 
     sinkOnEvent.connect<&GameLogic::slots_on_event>(*this);
@@ -57,21 +56,9 @@ game::GameLogic::GameLogic(ThePURGE &game) :
     sinkCastSpell.connect<&GameLogic::slots_cast_spell>(*this);
 
     sinkGetKilled.connect<&GameLogic::slots_kill_entity>(*this);
+    sinkDamageTaken.connect<&GameLogic::slots_damage_taken>(*this);
+
     sinkOnFloorChange.connect<&GameLogic::slots_change_floor>(*this);
-}
-
-auto game::GameLogic::slots_move([[maybe_unused]] entt::registry &world, entt::entity &player, const Direction &dir, bool is_pressed)
-    -> void
-{
-    auto spd = world.get<Speed>(player).speed;
-
-    switch (dir) {
-    case Direction::UP: world.get<ControllerAxis>(player).movement.y = is_pressed ? spd : 0.0f; break;
-    case Direction::DOWN: world.get<ControllerAxis>(player).movement.y = is_pressed ? -spd : 0.0f; break;
-    case Direction::RIGHT: world.get<ControllerAxis>(player).movement.x = is_pressed ? spd : 0.0f; break;
-    case Direction::LEFT: world.get<ControllerAxis>(player).movement.x = is_pressed ? -spd : 0.0f; break;
-    default: break;
-    }
 }
 
 auto game::GameLogic::slots_game_start(entt::registry &world) -> void
@@ -102,57 +89,6 @@ auto game::GameLogic::slots_game_start(entt::registry &world) -> void
     m_game.getCamera().setViewportSize(glm::vec2(25, 17));
 
     onFloorChange.publish(world);
-}
-
-auto game::GameLogic::slots_apply_classes(entt::registry &world, entt::entity player, const Class &newClass) -> void
-{
-    static auto holder = engine::Core::Holder{};
-
-    // world.get<AttackDamage>(player).damage = newClass.damage;
-    world.get<Speed>(player).speed = newClass.speed;
-    world.get<engine::d2::HitboxSolid>(player) = newClass.hitbox;
-
-    auto &health = world.get<Health>(player);
-
-    health.current += newClass.health;
-    health.max += newClass.health;
-    world.get<Classes>(player).ids.push_back(newClass.name);
-
-
-    // TODO: actual spell selection ?
-
-    for (const auto &spell : newClass.spells)
-        for (auto &slot : world.get<SpellSlots>(player).spells) {
-            if (slot.has_value()) continue;
-            slot = m_game.dbSpells().instantiate(spell);
-            break;
-        }
-
-    world.replace<engine::Spritesheet>(
-        player, engine::Spritesheet::from_json(holder.instance->settings().data_folder + newClass.assetGraphPath));
-
-    // Doesn't really matter, will be overridden by correct one soon enough. Prevent segfault of accessing inexistant
-    // "default" animation sp.current_animation = "idle_right";
-
-    engine::DrawableFactory::fix_spritesheet(world, player, "idle_right");
-}
-
-auto game::GameLogic::slots_purchase_classes(entt::registry &world, entt::entity player, const Class &newClass) -> void
-{
-    world.get<SkillPoint>(player).count -= newClass.cost;
-}
-
-auto game::GameLogic::slots_level_up(entt::registry &world, entt::entity entity) -> void
-{
-    world.get<SkillPoint>(entity).count++;
-
-    auto &level = world.get<Level>(entity);
-    level.current_xp -= level.xp_require;
-    level.xp_require = static_cast<std::uint32_t>(std::ceil(level.xp_require * 1.2));
-    level.current_level++;
-
-    auto pos = world.get<engine::d3::Position>(entity);
-    ParticuleFactory::create<Particule::POSITIVE>(world, {pos.x, pos.y}, {255, 255, 0});
 }
 
 auto game::GameLogic::slots_on_event(entt::registry &world, const engine::Event &e) -> void
@@ -231,28 +167,21 @@ auto game::GameLogic::slots_on_event(entt::registry &world, const engine::Event 
                 case engine::Joystick::LSX:
                 case engine::Joystick::LSY: {
                     auto joystick = holder.instance->getJoystick(joy.source.id);
-                    auto &axis = world.get<ControllerAxis>(player).movement;
 
-                    glm::vec2 newVal((*joystick)->axes[engine::Joystick::LSX], -(*joystick)->axes[engine::Joystick::LSY]);
-
-                    if (glm::length(newVal) > ControllerAxis::kDeadzone)
-                        axis = newVal;
-                    else
-                        axis = glm::vec2(0, 0);
+                    const auto newVal =
+                        glm::vec2{(*joystick)->axes[engine::Joystick::LSX], -(*joystick)->axes[engine::Joystick::LSY]};
+                    world.get<ControllerAxis>(player).movement =
+                        glm::length(newVal) > ControllerAxis::kDeadzone ? newVal : glm::vec2{0, 0};
 
                 } break;
                 case engine::Joystick::RSX:
                 case engine::Joystick::RSY: {
                     auto joystick = holder.instance->getJoystick(joy.source.id);
 
-                    auto &axis = world.get<ControllerAxis>(player).aiming;
-
-                    glm::vec2 newVal((*joystick)->axes[engine::Joystick::RSX], -(*joystick)->axes[engine::Joystick::RSY]);
-
-                    if (glm::length(newVal) > ControllerAxis::kDeadzone)
-                        axis = newVal;
-                    else
-                        axis = glm::vec2(0, 0);
+                    const auto newVal =
+                        glm::vec2{(*joystick)->axes[engine::Joystick::RSX], -(*joystick)->axes[engine::Joystick::RSY]};
+                    world.get<ControllerAxis>(player).aiming =
+                        glm::length(newVal) > ControllerAxis::kDeadzone ? newVal : glm::vec2{0, 0};
 
                 } break;
                 default: break;
@@ -275,61 +204,6 @@ auto game::GameLogic::slots_on_event(entt::registry &world, const engine::Event 
             [&](auto) {},
         },
         e);
-}
-
-auto game::GameLogic::slots_update_player_movement(entt::registry &world, [[maybe_unused]] const engine::TimeElapsed &dt)
-    -> void
-{
-    auto player = m_game.player;
-
-    auto &vel = world.get<engine::d2::Velocity>(player);
-    const auto &axis = world.get<ControllerAxis>(player);
-    auto spd = world.get<Speed>(player).speed;
-
-    vel.x = axis.movement.x * spd;
-    vel.y = axis.movement.y * spd;
-}
-
-auto game::GameLogic::slots_update_ai_movement(entt::registry &world, [[maybe_unused]] const engine::TimeElapsed &dt) -> void
-{
-    const auto pursue = [&world](entt::entity entity, entt::entity target, engine::d2::Velocity &out) {
-        const auto &pos = world.get<engine::d3::Position>(entity);
-        const auto &view_range = world.get<ViewRange>(entity);
-
-        const auto &target_pos = world.get<engine::d3::Position>(target);
-        const auto diff = glm::vec2{target_pos.x - pos.x, target_pos.y - pos.y};
-
-        if (glm::length(diff) > view_range.range) return false;
-
-        for (float i = 0.0f; i != 10.0f; i++) {
-            const auto in_between =
-                glm::vec2{static_cast<float>(pos.x) + i * diff.x / 10.0f, static_cast<float>(pos.y) + i * diff.y / 10.0f};
-            for (const auto &wall : world.view<entt::tag<"wall"_hs>>()) {
-                const auto &wall_pos = world.get<engine::d3::Position>(wall);
-                const auto &wall_hitbox = world.get<engine::d2::HitboxSolid>(wall);
-
-                if (engine::d2::overlapped<engine::d2::WITHOUT_EDGE>(
-                        engine::d2::HitboxSolid{0.01, 0.01},
-                        engine::d3::Position{in_between.x, in_between.y, 0.0f},
-                        wall_hitbox,
-                        wall_pos)) {
-                    return false;
-                }
-            }
-        }
-
-        const auto result = glm::normalize(diff) * 7.f;
-        out = {result.x, result.y};
-
-        return true;
-    };
-
-    for (auto &i : world.view<entt::tag<"enemy"_hs>>()) {
-        if (world.has<engine::Spritesheet>(i) && world.get<engine::Spritesheet>(i).current_animation == "death")
-            continue;
-        auto &vel = world.get<engine::d2::Velocity>(i);
-        pursue(i, m_game.player, vel);
-    }
 }
 
 auto game::GameLogic::slots_update_cooldown(entt::registry &world, const engine::TimeElapsed &dt) -> void
@@ -370,117 +244,6 @@ auto game::GameLogic::slots_update_effect(entt::registry &world, const engine::T
             effect.is_in_effect = false;
         }
     });
-}
-
-auto game::GameLogic::slots_update_ai_attack(entt::registry &world, [[maybe_unused]] const engine::TimeElapsed &dt) -> void
-{
-    for (auto enemy : world.view<entt::tag<"enemy"_hs>, engine::d3::Position, AttackRange>()) {
-        // TODO: Add brain to AI. current strategy : spam every spell towards the player
-
-        if (world.has<engine::Spritesheet>(enemy) && world.get<engine::Spritesheet>(enemy).current_animation == "death")
-            continue;
-
-        for (auto &spell : world.get<SpellSlots>(enemy).spells) {
-            if (!spell.has_value()) continue;
-
-            const auto &selfPosition = world.get<engine::d3::Position>(enemy);
-            const auto &targetPosition = world.get<engine::d3::Position>(m_game.player);
-
-            const glm::vec2 diff = {targetPosition.x - selfPosition.x, targetPosition.y - selfPosition.y};
-
-            auto &attack_range = world.get<AttackRange>(enemy);
-
-            if (glm::length(diff) <= attack_range.range) {
-                onSpellCast.publish(world, enemy, {diff.x, diff.y}, spell.value());
-            }
-        }
-    }
-}
-
-auto game::GameLogic::slots_check_collision(entt::registry &world, [[maybe_unused]] const engine::TimeElapsed &dt) -> void
-{
-    static auto holder = engine::Core::Holder{};
-
-    for (auto &spell : world.view<entt::tag<"spell"_hs>>()) {
-        const auto &spell_pos = world.get<engine::d3::Position>(spell);
-        const auto &spell_box = world.get<engine::d2::HitboxFloat>(spell);
-
-        for (auto &wall : world.view<entt::tag<"wall"_hs>>()) {
-            const auto &wall_pos = world.get<engine::d3::Position>(wall);
-            const auto &wall_box = world.get<engine::d2::HitboxSolid>(wall);
-
-            if (engine::d2::overlapped<engine::d2::WITH_EDGE>(spell_box, spell_pos, wall_box, wall_pos)) {
-                world.destroy(spell);
-                break;
-            }
-        }
-    }
-
-    const auto apply_damage = [this, &world](auto &entity, auto &spell, auto &spell_hitbox, auto &spell_pos, auto &source) {
-        if (world.has<engine::Spritesheet>(entity) && world.get<engine::Spritesheet>(entity).current_animation == "death")
-            return;
-
-        auto &entity_pos = world.get<engine::d3::Position>(entity);
-        auto &entity_hitbox = world.get<engine::d2::HitboxSolid>(entity);
-
-
-        if (engine::d2::overlapped<engine::d2::WITHOUT_EDGE>(entity_hitbox, entity_pos, spell_hitbox, spell_pos)) {
-            auto &entity_health = world.get<Health>(entity);
-            auto &spell_damage = world.get<AttackDamage>(spell);
-
-            // todo : publish entity_took_damage instead
-            entity_health.current -= spell_damage.damage;
-
-            if (world.has<entt::tag<"player"_hs>>(entity)) {
-                holder.instance->setScreenshake(true, 350ms);
-                ParticuleFactory::create<Particule::HITMARKER>(
-                    world, {(spell_pos.x + entity_pos.x) / 2.0, (entity_pos.y + spell_pos.y) / 2.0}, {255,0,0});
-            } else {
-                ParticuleFactory::create<Particule::HITMARKER>(
-                    world, {(spell_pos.x + entity_pos.x) / 2.0, (entity_pos.y + spell_pos.y) / 2.0}, {0, 0, 0});
-            }
-
-            holder.instance->getAudioManager()
-                .getSound(holder.instance->settings().data_folder + "sounds/fire_hit.wav")
-                ->play();
-            world.destroy(spell);
-            if (entity_health.current <= 0.0f) { onEntityKilled.publish(world, entity, source); }
-        }
-    };
-
-    for (auto &spell : world.view<entt::tag<"spell"_hs>>()) {
-        const auto &source = world.get<engine::Source>(spell).source;
-        if (!world.valid(source)) return;
-
-        auto &spell_pos = world.get<engine::d3::Position>(spell);
-        const auto &hitbox = world.get<engine::d2::HitboxFloat>(spell);
-
-        if (world.has<entt::tag<"player"_hs>>(source)) {
-            for (auto &enemy : world.view<entt::tag<"enemy"_hs>>()) {
-                if (world.valid(spell)) apply_damage(enemy, spell, hitbox, spell_pos, source);
-            }
-        } else {
-            for (auto &player : world.view<entt::tag<"player"_hs>>()) {
-                if (world.valid(spell)) apply_damage(player, spell, hitbox, spell_pos, source);
-            }
-        }
-    }
-
-    world.view<KeyPicker, engine::d2::HitboxSolid, engine::d3::Position>().each(
-        [&](KeyPicker &keypicker, const engine::d2::HitboxSolid &pickerhitbox, const engine::d3::Position &pickerPos) {
-            if (keypicker.hasKey) return;
-
-            for (auto &key : world.view<entt::tag<"key"_hs>>()) {
-                auto &keyHitbox = world.get<engine::d2::HitboxFloat>(key);
-                auto &keyPos = world.get<engine::d3::Position>(key);
-
-                if (engine::d2::overlapped<engine::d2::WITH_EDGE>(pickerhitbox, pickerPos, keyHitbox, keyPos)) {
-                    keypicker.hasKey = true;
-
-                    world.destroy(key);
-                }
-            }
-        });
 }
 
 auto game::GameLogic::slots_update_particle(
@@ -606,56 +369,6 @@ auto game::GameLogic::slots_update_player_sigh(entt::registry &world, const engi
     sightScale.y = scale * kSightScaleMultiplier;
 }
 
-auto game::GameLogic::slots_kill_entity(entt::registry &world, entt::entity killed, entt::entity killer) -> void
-{
-    static auto holder = engine::Core::Holder{};
-
-    engine::DrawableFactory::fix_spritesheet(world, killed, "death");
-    const auto &animation = world.get<engine::Spritesheet>(killed).animations["death"];
-
-    world.emplace<engine::Lifetime>(killed, std::chrono::milliseconds(animation.frames.size() * animation.cooldown));
-    world.get<engine::d2::Velocity>(killed) = {0.0, 0.0};
-
-    world.remove<engine::d2::HitboxSolid>(killed);
-
-    if (world.has<entt::tag<"player"_hs>>(killed)) {
-        holder.instance->getAudioManager()
-            .getSound(holder.instance->settings().data_folder + "sounds/player_death.wav")
-            ->play();
-
-        m_game.setMenu(std::make_unique<menu::GameOver>());
-
-    } else if (world.has<entt::tag<"enemy"_hs>>(killed)) {
-        // TODO: actual random utilities
-        bool lazyDevCoinflip = static_cast<std::uint32_t>(killed) % 2;
-        holder.instance->getAudioManager()
-            .getSound(
-                lazyDevCoinflip ? holder.instance->settings().data_folder + "sounds/death_01.wav"
-                                : holder.instance->settings().data_folder + "sounds/death_02.wav")
-            ->play();
-
-        if (world.has<entt::tag<"player"_hs>>(killer)) { addXp(world, killer, world.get<Experience>(killed).xp); }
-
-        if (world.has<entt::tag<"boss"_hs>>(killed)) {
-            auto pos = world.get<engine::d3::Position>(killed);
-            EntityFactory::create<EntityFactory::KEY>(m_game, world, {pos.x, pos.y}, {1.0, 1.0});
-            holder.instance->getAudioManager()
-                .getSound(holder.instance->settings().data_folder + "sounds/boss_death.wav")
-                ->play();
-        }
-    }
-}
-
-auto game::GameLogic::slots_cast_spell(entt::registry &world, entt::entity caster, const glm::dvec2 &direction, Spell &spell)
-    -> void
-{
-    if (!spell.cd.is_in_cooldown) {
-        SpellFactory::create(world, caster, glm::normalize(direction), m_game.dbSpells().db.at(std::string{spell.id}));
-        spell.cd.remaining_cooldown = spell.cd.cooldown;
-        spell.cd.is_in_cooldown = true;
-    }
-}
-
 auto game::GameLogic::slots_change_floor(entt::registry &world) -> void
 {
     world.view<entt::tag<"terrain"_hs>>().each([&](auto &e) { world.destroy(e); });
@@ -674,13 +387,4 @@ auto game::GameLogic::slots_change_floor(entt::registry &world) -> void
         pos.x = data.spawn.x + data.spawn.w * 0.5;
         pos.y = data.spawn.y + data.spawn.h * 0.5;
     }
-}
-
-auto game::GameLogic::addXp(entt::registry &world, entt::entity player, std::uint32_t xp) -> void
-{
-    auto &level = world.get<Level>(player);
-
-    level.current_xp += xp;
-
-    while (level.current_xp >= level.xp_require) { onPlayerLevelUp.publish(world, player); }
 }
